shuffle around some files
