+++ packages/frontend-shared/cypress/support/common.ts
@@ -17,18 +17,6 @@ export function addVueCommand () {
   })
 }
 
-export function logInternal<T> (name: string | Partial<Cypress.LogConfig>, cb: (log: Cypress.Log | undefined) => Cypress.Chainable<T>, opts: Partial<Cypress.Loggable> = {}): Cypress.Chainable<T> {
-  const _log = typeof name === 'string'
-    ? Cypress.log({ name, message: '' })
-    : Cypress.log(name)
-
-  return cb(_log).then<T>((val) => {
-    _log?.end()
-
-    return val
-  })
-}
-
 declare global {
   namespace Cypress {
     interface Chainable {
